feat: Add lottie-react package
This commit adds the "lottie-react" package to the project dependencies. It is added with the version "^2.4.0". This package allows for the use of Lottie animations in the project.

The "lottie-react" package is added to both the "package.json" file and the "pnpm-lock.yaml" file. Additionally, a lottie transition animation file is added to the "src/lottie" directory.

In the "Assessment.tsx" file, the "Lottie" component is imported from "lottie-react" and used to display a Lottie animation during the assessment process. The animation data is sourced from the newly